Extracted a view view components of the settings page into their own files.
